feat: add extra repository

+++ src/modules/categories/domain/use-cases/delete-category-usecase/delete-category.usecase.ts
@@ -1,11 +1,16 @@
 import { CategoryRepository } from 'src/modules/categories/ports/repositories/category.repository';
+import { CustomerNotFoundExecption } from 'src/modules/common/core/exceptions/not-found.exception';
 
 export class DeleteCategoryUseCase {
   constructor(private readonly categoryRepository: CategoryRepository) {}
 
   async execute(id: string): Promise<void> {
-    const category = await this.categoryRepository.findOne(id);
+    const category = await this.categoryRepository.findId(id);
 
+    if(!category) {
+      throw new CustomerNotFoundExecption("Category not found")
+    }
+    
     await this.categoryRepository.delete(category.id);
   }
 }
